+++ GTGen.h
@@ -1,6 +1,6 @@
 #ifndef GTGEN_H
 #define GTGEN_H
 
-int generate_solutions_gt(int N);
+unsigned long long generate_solutions_gt(int N);
 
 #endif /* GTGEN_H */
